Imported Upstream version 3.0.0~beta1 Conflicts: 	.gitignore

+++ lib/map.h
@@ -38,8 +38,6 @@
  * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN
  * AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING
  * OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
- *
- * $Id: map.h,v 1.14 2010/01/06 17:01:46 murch Exp $
  */
 
 #ifndef INCLUDED_MAP_H
@@ -54,21 +52,21 @@ extern const char *map_method_desc;
  * fd is the file descriptor which is to be mapped
  * onceonly is set to be nonzero if you do not intend to ever refresh the map
  * base and len are output parameters that receive the address and length
- * 	of the map once it is created.  NOTE: *len should be zero the first
- * 	time map_refresh() is called to force the initial mapping
+ *      of the map once it is created.  NOTE: *len should be zero the first
+ *      time map_refresh() is called to force the initial mapping
  * newlen is set to the size of the file, or MAP_UNKNOWN_LEN to have the
- * 	mapping facility compute it for you.
+ *      mapping facility compute it for you.
  * name and mboxname are used for logging purposes, name is the name
- * 	of the file, and shouldn't be NULL, while mboxname is the name
- * 	of the applicable mailbox (if any), and may be NULL
+ *      of the file, and shouldn't be NULL, while mboxname is the name
+ *      of the applicable mailbox (if any), and may be NULL
  */
 extern void map_refresh(int fd, int onceonly, const char **base,
-			unsigned long *len, unsigned long newlen,
-			const char *name, const char *mboxname);
+                        size_t *len, size_t newlen,
+                        const char *name, const char *mboxname);
 
 /* map_free will free a memory map allocated by map_refresh
  *
  * base and len are the same values that were passed to map_refresh */
-extern void map_free(const char **base, unsigned long *len);
+extern void map_free(const char **base, size_t *len);
 
 #endif /* INCLUDED_MAP_H */
